Release version v1.0.0 (#427)
* relationship fixes

* add onDelete cascade to team relationships in entity models

* implement session token cleanup and logging; refactor token expiration handling

* add meta tags for SEO optimization and enhance page metadata for Quick Learn

* some fixes

* add global search parameters interface and update related services

* refactor(auth, course, users): streamline service methods and enhance frontend URL handling

* feat(leaderboard): enhance leaderboard data retrieval with team_id support

* community fixes and import fixes

* deps: bump @nx/node from 20.4.6 to 20.6.2

Bumps [@nx/node](https://github.com/nrwl/nx/tree/HEAD/packages/node) from 20.4.6 to 20.6.2.
- [Release notes](https://github.com/nrwl/nx/releases)
- [Commits](https://github.com/nrwl/nx/commits/20.6.2/packages/node)

---
updated-dependencies:
- dependency-name: "@nx/node"
  dependency-type: direct:development
  update-type: version-update:semver-minor
...


* feat(leaderboard, course): add team_id support and enhance error handling

* feat(skills, roadmap): add team_id support and improve error handling in skill and roadmap operations

* feat(next.config): enable unoptimized images for improved performance

* course category edit/update issue fix

* Fix: issues with the updation of the course category

* feat(auth): update cookie handling for refresh token and user role

* community course and member details loading issue fix

* chore(.gitignore): add entry to ignore cursor folders

* feat(manifest): add PWA support with icons and metadata configuration

* fix: update loading state handling in TeamMemberDetails component and ignore cursor settings

* checkbox, loader fix

* fix: linting and unflag lesson issue

* feat: add cascading delete for user foreign key in leaderboard entities

* draft fix

* Chore: updated the package-lock file

* breadcrumb fix

* chore: update pirates package to version 4.0.7 in package-lock.json

* refactor: remove console logs and improve condition checks in services and components

---------

+++ apps/quick-learn-backend/src/routes/course-category/course-category.service.ts
@@ -2,7 +2,7 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateCourseCategoryDto } from './dto/create-course-category.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ILike, Repository } from 'typeorm';
-import { CourseCategoryEntity } from '@src/entities';
+import { CourseCategoryEntity, UserEntity } from '@src/entities';
 import { BasicCrudService } from '@src/common/services';
 import { UpdateCourseCategoryDto } from './dto/update-course-category.dto';
 import { en } from '@src/lang/en';
@@ -16,68 +16,66 @@ export class CourseCategoryService extends BasicCrudService<CourseCategoryEntity
     super(courseCategoryRepository);
   }
 
-  async create(createCourseCategoryDto: CreateCourseCategoryDto) {
+  async createCourseCategory(
+    createCourseCategoryDto: CreateCourseCategoryDto,
+    user: UserEntity,
+  ) {
     const foundCourseCategory = await this.repository.count({
-      where: { name: ILike(createCourseCategoryDto.name) },
+      where: {
+        name: ILike(createCourseCategoryDto.name),
+        team_id: user.team_id,
+      },
     });
     if (foundCourseCategory) {
       throw new BadRequestException('Course Category already exists');
     }
-    const courseCategory = this.repository.create(createCourseCategoryDto);
+    const courseCategory = this.repository.create({
+      ...createCourseCategoryDto,
+      team_id: user.team_id,
+    });
     return await this.repository.save(courseCategory);
   }
 
-  async createCourseCategory(
+  async updateCourseCategory(
     id: number,
-    createCourseCategoryDto: UpdateCourseCategoryDto,
+    updateCourseCategoryDto: UpdateCourseCategoryDto,
+    team_id: number,
   ) {
-    const courseCategory = await this.get({ id });
+    const courseCategory = await this.get({ id, team_id });
+
+    if (!courseCategory) {
+      throw new BadRequestException(en.InvalidCourseCategory);
+    }
+
     const foundCourseCategory = await this.get({
-      name: ILike(createCourseCategoryDto.name),
+      name: ILike(updateCourseCategoryDto.name),
+      team_id,
     });
+
     if (foundCourseCategory && foundCourseCategory.id !== courseCategory.id) {
       throw new BadRequestException('Course Category already exists');
     }
-    return await this.update({ id }, createCourseCategoryDto);
+
+    return await this.update({ id }, updateCourseCategoryDto);
   }
 
-  async getAllCourseCategoriesWithLessonsCount(): Promise<
-    CourseCategoryEntity[]
-  > {
-    return await this.repository
-      .createQueryBuilder('course_category')
-      .leftJoinAndSelect(
-        'course_category.courses',
-        'courses',
-        'courses.archived = :archived',
-        {
-          archived: false,
-        },
-      )
-      .leftJoin('courses.lessons', 'lessons')
-      .loadRelationCountAndMap(
-        'courses.lessons_count',
-        'courses.lessons',
-        'lessons',
-        (qb) =>
-          qb.andWhere('lessons.archived = :archived', { archived: false }),
-      )
-      .orderBy('courses.created_at', 'DESC')
-      .addOrderBy('course_category.created_at', 'DESC')
-      .getMany();
-  }
-
-  async deleteCourseCategory(id: number): Promise<void> {
+  async deleteCourseCategory(id: number, user: number): Promise<void> {
     const courseCategory = await this.get({ id }, ['courses']);
+    if (courseCategory.team_id !== user) {
+      throw new BadRequestException(
+        'You do not have permission to delete this course category',
+      );
+    }
     if (courseCategory.courses.length > 0) {
       throw new BadRequestException(en.courseCategriesHasData);
     }
     await this.repository.delete({ id });
   }
 
-  async getCourseCategoriesWithCourses() {
+  async getCourseCategoriesWithCourses({ team_id }: UserEntity) {
     return await this.repository
       .createQueryBuilder('course_category')
+      .where('course_category.team_id = :teamId', { teamId: team_id })
       .leftJoinAndSelect(
         'course_category.courses',
         'courses',
@@ -91,7 +89,9 @@ export class CourseCategoryService extends BasicCrudService<CourseCategoryEntity
         'courses.lessons',
         'lessons',
         (qb) =>
-          qb.andWhere('lessons.archived = :archived', { archived: false }),
+          qb.andWhere('lessons.archived = :archived', {
+            archived: false,
+          }),
       )
       .leftJoin('courses.roadmaps', 'roadmaps')
       .loadRelationCountAndMap(
